pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "requests>=2.32.4",
    "streamlit-authenticator>=0.4.2",
    "streamlit>=1.46.1",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "sqlalchemy>=2.0.41",
    "bcrypt>=4.3.0",
    "psycopg2-binary>=2.9.10",
]
